+++ app/sprinkles/core/config/default.php
@@ -164,7 +164,7 @@
                     'port'              => isset($_SERVER['SERVER_PORT']) ? (int) $_SERVER['SERVER_PORT'] : null,
                     'path'              => isset($_SERVER['SCRIPT_NAME']) ? trim(dirname($_SERVER['SCRIPT_NAME']), '/\\') : ''
                 ],
-                'author'            => 'http://www.userfrosting.com',
+                'author'            => 'https://www.userfrosting.com',
                 'publisher'         => ''
             ]
         ],
@@ -172,6 +172,8 @@
             'timezone' => 'America/New_York',
             'error_reporting' => E_ALL,  // Development - report all errors and suggestions
             'display_errors'  => 'true',
-            'log_errors'      => 'false'
+            'log_errors'      => 'false',
+            // Let PHP itself render errors natively.  Useful if a fatal error is raised in our custom shutdown handler.
+            'display_errors_native' => 'false'
         ]
     ];